pyproject.toml
[project]
name = "wood-carving"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cairosvg>=2.8.2",
    "colorlog>=6.10.1",
    "geopandas>=1.1.1",
    "lxml>=6.0.1",
    "matplotlib>=3.10.6",
    "osmium>=4.1.1",
    "osmnx>=2.0.6",
    "pyyaml>=6.0.3",
    "shapely>=2.1.1",
    "svgwrite>=1.4.3",
]
